+++ internal/app/app.go
@@ -8,6 +8,7 @@ import (
 	"github.com/mih-kopylov/our-spb-bot/internal/queue"
 	"github.com/mih-kopylov/our-spb-bot/internal/spb"
 	"github.com/mih-kopylov/our-spb-bot/internal/state"
+	"github.com/mih-kopylov/our-spb-bot/internal/storage"
 	"github.com/samber/lo"
 	"github.com/sirupsen/logrus"
 )
@@ -15,6 +16,7 @@ import (
 func RunApplication() error {
 	logrus.SetLevel(logrus.DebugLevel)
 	conf := config.RegisterBean()
+	storage.RegisterBean(conf)
 	state.RegisterBean()
 	category.RegisterBean()
 	bot.RegisterApiBean(conf)