+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Firebase to store user state and messages queue
+
 ## [0.2.0] - 2023-03-06
 
 ### Added